Refactor NFC tag filter method and remove unused code
