+++ backend/digidex/settings/development.py
@@ -49,7 +49,3 @@ LOGGING = {
         },
     },
 }
-
-
-# Wagtail-ntag Configuration (BASE)
-NFC_TAG_FILTER_METHOD = 'uid'
