+++ backend/digidex/settings/production.py
@@ -154,8 +154,3 @@ LOGGING = {
         },
     },
 }
-
-# ------------------------------------------------------------------------
-# Wagtail-ntag Configuration (PROD)
-# ------------------------------------------------------------------------
-NFC_TAG_FILTER_METHOD = 'uid_counter'
