+++ backend/ntags/api.py
@@ -1,11 +1,62 @@
-from django.urls import path, include
 from rest_framework.routers import DefaultRouter
+from rest_framework import status, viewsets, permissions
+from rest_framework_simplejwt.authentication import JWTAuthentication
+from rest_framework.response import Response
 
-from .viewsets import NFCTagViewSet
+from .models import NFCTag
+from .serializers import NFCTagSerializer
+
+
+class NFCTagAPIViewSet(viewsets.ModelViewSet):
+    """
+    A viewset for viewing, creating, and linking NFC tags.
+    """
+    queryset = NFCTag.objects.all()
+    serializer_class = NFCTagSerializer
+    permission_classes = [permissions.IsAuthenticatedOrReadOnly]
+    authentication_classes = [JWTAuthentication]
+    lookup_field = 'serial_number'
+
+    def get_queryset(self):
+        if self.request.user.is_superuser:
+            return NFCTag.objects.all()
+        return NFCTag.objects.filter(
+            user=self.request.user,
+            active=True
+        )
+
+    def create(self, request, *args, **kwargs):
+        """
+        Create a new NFC tag with the provided serial number.
+        """
+        serial_number = request.data.get('serial_number')
+
+        if not serial_number:
+            return Response({"error": "Serial Number not provided."}, status=status.HTTP_400_BAD_REQUEST)
+
+        ntag, created = NFCTag.objects.update_or_create(
+            serial_number=serial_number
+        )
+
+        serializer = self.get_serializer(ntag, context={'request': request})
+        return Response(serializer.data, status=status.HTTP_201_CREATED if created else status.HTTP_200_OK)
+
+    def destroy(self, request, *args, **kwargs):
+        """
+        Perform actual deletion if the user is a superuser, otherwise set
+        the NFC tag's active status to False.
+        """
+        instance = self.get_object()
+
+        if request.user.is_superuser:
+            # Perform the usual delete if the user is a superuser
+            return super().destroy(request, *args, **kwargs)
+        else:
+            # For regular users, set 'active' to False instead of deleting
+            instance.active = False
+            instance.save()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+
 
 router = DefaultRouter()
-router.register(r'ntags', NFCTagViewSet, basename='ntag')
-
-urlpatterns = [
-    path('api/', include(router.urls)),
-]
+router.register(r'ntags', NFCTagAPIViewSet, basename='ntag')  # noqa
